fix types

+++ anikanov.vladislav/P3/REPDGT.cpp
@@ -6,7 +6,7 @@
 int hasDuplicateDigits(const char *str)
 {
   int entrance[10] = {0};
-  for (int i = 0; i < strlen(str); ++i) {
+  for (size_t i = 0; i < strlen(str); ++i) {
     if (std::isdigit(str[i])){
       entrance[str[i] - '0']++;
     }
